Add styling for the navbarsection
