+++ src/App.css
@@ -1,15 +1,21 @@
-.App {
-  text-align: center;
+:root {
+  --pale-grey: #f5f6fa;
+  --white-three: rgba(0, 0, 0, 0.05);
 }
 
-.App-header {
-  background-color: #f4f5f7;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(6px + 2vmin);
-  color: rgb(13, 12, 12);
-  list-style-type: none;
+.App {
+  width: 100rem;
+  height: 89.62rem;
+  padding: 9.313rem 3.75rem 7.688rem 6.25rem;
+  background-color: var(--pale-grey);
+}
+
+.panel-bg {
+  width: 87.5rem;
+  height: 68.563rem;
+  margin: 0 2.5rem 4.063rem 0;
+  padding: 0 0 49.625rem;
+  border-radius: 4px;
+  box-shadow: 0 0 60px 0 rgba(0, 0, 0, 0.05);
+  background-color: var(--white-three);
 }
