+++ src/components/Books.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import AddBook from './AddBook';
-import BooksList from './BooksList';
+import BooksList from './BookList';
 
 const Books = () => (
   <div className="booksWrapper">
